wip: add modal parent | handle back | reintegrate send
